+++ server/src/Application/Dtos/Incoming/Outgoing/CourseDto.cs
@@ -1,3 +1,5 @@
+using StudyZen.Domain.Entities;
+
 namespace StudyZen.Application.Dtos;
 
 public class CourseDto
@@ -5,4 +7,14 @@ public class CourseDto
     public int Id { get; set; }
     public string Name { get; set; }
     public string Description { get; set; }
+
+    public static CourseDto ToDto(Course course)
+    {
+        return new CourseDto
+        {
+            Id = course.Id,
+            Name = course.Name,
+            Description = course.Description
+        };
+    }
 }
