+++ server/src/Application/Dtos/Incoming/Outgoing/FlashcardDto.cs
@@ -1,3 +1,5 @@
+using StudyZen.Domain.Entities;
+
 namespace StudyZen.Application.Dtos;
 
 public class FlashcardDto 
@@ -6,5 +8,17 @@ public class FlashcardDto
     public int FlashcardSetId { get; set; }
     public string Question { get; set; }
     public string Answer { get; set; }
+
+    public static FlashcardDto ToDto(Flashcard flashcard)
+    {
+        return new FlashcardDto
+        {
+            Id = flashcard.Id,
+            FlashcardSetId = flashcard.FlashcardSetId,
+            Question = flashcard.Question,
+            Answer = flashcard.Answer
+        };
+    }
+
 
 }
