+++ server/src/Application/Dtos/Incoming/Outgoing/FlashcardSetDto.cs
@@ -1,4 +1,5 @@
 using StudyZen.Domain.Enums;
+using StudyZen.Domain.Entities;
 
 namespace StudyZen.Application.Dtos;
 
@@ -8,4 +9,15 @@ public class FlashcardSetDto
     public int? LectureId { get; set; }
     public string Name { get; set; }
     public Color Color { get; set; }
+
+    public static FlashcardSetDto ToDto(FlashcardSet flashcardSet)
+    {
+        return new FlashcardSetDto
+        {
+            Id = flashcardSet.Id,
+            LectureId = flashcardSet.LectureId,
+            Name = flashcardSet.Name,
+            Color = flashcardSet.Color
+        };
+    }
 }
