+++ server/src/Application/Dtos/Incoming/Outgoing/LectureDto.cs
@@ -1,3 +1,5 @@
+using StudyZen.Domain.Entities;
+
 namespace StudyZen.Application.Dtos;
 
 public class LectureDto
@@ -6,4 +8,15 @@ public class LectureDto
     public int CourseId { get; set; }
     public string Name { get; set; }
     public string Content { get; set; }
+
+    public static LectureDto ToDto(Lecture lecture)
+    {
+        return new LectureDto
+        {
+            Id = lecture.Id,
+            CourseId = lecture.CourseId,
+            Name = lecture.Name,
+            Content = lecture.Content
+        };
+    }
 }
